Simplify `all_asset_variants_uploaded?` method
Now that the logic defining `version :thumbnail, if: :pdf?` in
`AttachmentUploader` has been removed, there is no attachment
type that has anything other than an `:original` version. We
therefore know that `required_variants` is always just `:original`
and can simplify the method.

I did consider renaming the method to simply `uploaded?` or
`original_uploaded?`, but a method called
`all_asset_variants_uploaded?` is defined in 8 different places,
and called in almost a hundred places, so it's difficult to know
which variation of the method is being called in which place (and
therefore which method calls to rename). Safer to just leave it.

With the method simplified, we can remove the test (since it
would no longer work).