Add rake task to update missing replacement links
We've identified a bug where if the user replaces an attachments on a new draft,
without having saved the edition draft first, the supposedly replaced original
asset remains live, because it fails to redirect to its replacement.

This rake task figures out the last AttachmentData in the chain, and reruns
the updater on the original AttachmentData with the replacement_id set to the
last AttachmentData's asset_manager_id. It fetches the ID based on variant
value, sending the correct `replacement_id` field to `asset-manager`.

This is only a data patch solution.
A fix for the issue will be implemented separately.