Add rake task to update missing replacement links
We've identified a bug where if the user replaces an attachments on a new draft,
without having saved the edition draft first, the supposedly replaced original
asset remains live, because it fails to redirect to its replacement.

Rerunning the updater on the original attachment's AttachmentData should send
the correct `replacement_id` field to `asset-manager`, and fix the data.

This is only a data patch solution.
A fix for the issue will be implemented separately.