Filter out legacy assets when checking if all assets uploaded
If the asset doesn't respond to `variant`, it's a legacy asset
(PDF thumbnail) and we want to filter it out.

I did consider applying this filter more generally by overloading
the `assets` call as below, but that breaks a bunch of tests - so
any filtering needs to be applied on a per-case basis.

```
  def assets
    super.select { |asset| asset.variant.present? }
  end
```