Prevent navigation events from being tracked on tab controls
It would be preferable to use the role ARIA attribute to do this, but it's not present yet when this module is initialised.
This is because component modules are initialised after analytics and the tabs component is responsible for adding the role via javascript during its initialisation.
Analytics modules are initialised before components by GOV.UK Publishing components.