Add basics for editionable topical events
There's been an RFC accepted to editionablise all the remaining content types
that are currently published by Whitehall. As part of a gift week project we're
attempting to do that to topical events. We're unsure as to how far we'll get
and will be liasing with the Whitehall Tech team to ensure we don't do anything
too silly.

This sets up the core components needed for Topical events. We'll build on this
until we're as close to feature pairity as we can reasonably get in a week.

https://github.com/alphagov/govuk-rfcs/blob/main/rfc-177-edition-all-content.md