Use Query class in `Document::PaginatedTimeline`
We expect to be making some changes to this class in the future, and
it’s already quite complex. Seperating out the behaviour makes it easier
to follow and reason about.

I could break up the tests and make them a little more “unit-y” by
mocking out the behavaiour of the presenter and the query class, but I
think it’s useful to keep the tests as-is for now.