validate state for edition under review
during the review stage, we want to check the
scheduled values from the previous page are valid,
but not yet put the edition into a scheduled or
published state.

here we validate there are no errors by updating
the edition via the `assign_attributes` method which
does not save the changes to the db.