Remove dashboard filters for orphaned corporate info pages
These filters were added in 2439f413 to work around the fact that we had a number of orphaned corporate information pages in the Whitehall database.

We have since removed all of the orphaned corporate info pages. Unfortunately we did not use a data migration to do this so I can't link to the commit.

We also believe we have fixed the issue that was causing the orphaned pages to be created. I believe this was done during the move to make worldwide organisations editionable but there isn't really a specific commit to link to.

As a result of both the above, it should be safe to remove this code.