Convert all tables to use UTF8MB4 charset
We have seen an increasing number of errors recently relating to publishers attempting to use 4 byte characters in Whitehall.
This commit adds a migration which will convert all of the tables to use the UTF8MB4 charset.

The migration will need to be run out of hours as the tables will need to be locked during conversion.