Allow deletion of review reminders
Previously the only way to remove the review date from a document was to create a new edition and uncheck the set review date checkbox, then save the edition.

If the user publishes the new edition but neglects to remove the review reminder, they will receive another reminder even though they updated the content.

This commit makes it possible for the user to remove the reminder from the published edition without having to create another edition.