Remove existing form tracking
The existing form tracker will not meet the form tracking needs across Whitehall and the other publishing apps. Having one instance the form tracker is not useful and we will be beginning work next year to create a generic solution for form tracking. To avoid confusion we should existing form tracking for the time being.