add rake task to populate Edition titles with their Document titles
we want to move the title ownership to Editions,
because the title can change and we want to record
the different versions of it. This rake task will
populate the Edition's title column with the
current Document title - I have put some
safeguarding in to ensure we don't overwrite any
existing data.

Once we have merged this PR with changes
to how the title is assigned, and run the task
once, there should be no need to run it again.