ensure any new editions created have their title updated
we want to move the title field to Content Block
Editions rather than Content Block Documents.
We have created a rake task to populate the field
for old editions. This commit not makes sure that
any new Editions created will have the title field
populated. Once this PR and the rake task have
been done, we should be safe to remove the old
Document title and use the new Edition title
everywhere.

I've added some tests and checked around, and I
think that the creation of any Edition only
happens in one place - the Create Edition Service
- so that's where we make
the change.