WIP on refactoring lead orgs to use add another and get rid of weird orgs association
I had to give up on this because we can't use select with search in conjunction with the add another component. One would need to reinitialise the select with search component each time a new fieldset was added, which will add substantial complexity to our Javascript code as currently all the modules function very independently.

This is frustrating because being able to remove the home grown edition organisations association management code was looking quite feasible.