Return editor object when fetching host content
This calls `ContentBlockManager::HostContentItem::Editor.with_uuids`
with all the editor IDs when fetching a page of results, meaning we
only have to make one API call per page. I’ve kept
`last_edited_by_editor_id` in the response for now, but will remove this
in a later commit.