remove title from Content Block Document
and use the title on Content Block Editions.

In this big commit, we have to make a lot of
changes which depend on each other:

* Remove `title` data from Documents.
This is slightly more complicated than just
removing the column, because the
`content_id_alias` was relying on the Document's
title. To get around that, I've renamed the title
column to `sluggable_string`, as Whitehall do for
their Document. For the Content Block Manager,
this field will only be assigned when the Document
is first created, though it could be updated in
the future when the title is changed. I don't
think we are currently using the
`content_id_alias` for CBM on the publishing side,
but there may be work done to use it as a
human-readable ID in the future.

* assign the `sluggable_string`` whenever we create a
  new Document.

* no longer update the Document `title` from the
  form data, instead update the Edition `title`.

* Point Document’s title method to latest Edition
  - so we have a helper method to get the title
  straight from Doc.