Fix flaky content block updates test
Because [`HostContentUpdateEvent#is_for_newer_edition?`][1] relies on
the date the edition was superseded, sometimes the tests are too fast
for their own good and everything happens at the same time (or within
too small a window for Rails to notice), so the timeline entry shows
up as on the current edition. This is extremely unlikely to happen in
Real Life, so we add an artificial passage of time using Timecop to
give a more realistic gap between when then first edition was created
and when it was superseded. This should fix the flakiness we’ve seen.

[1]: https://github.com/alphagov/whitehall/blob/3971e93a2c7612d95e38c439c4bbf9752eaf5c2c/app/models/host_content_update_event.rb#L19