Set a parent_class for the error_summary
Out of the box, the error messages generate a link with slashes when a
class is namespaced, which causes issues when generating IDs for labels
to link to (as IDs should only contain alphanumeric characters,
underscores and slashes). We can get around this by passing a
`parent_class` argument to the error summary component, so errors link
predictably.