Stop validating change notes when saving edition drafts
We want to stop validating change notes when saving edition drafts because it can cause attachment replacement to fail. Attachment replacement will fail if the edition is invalid, and it is possible to replace an attachment before saving a new edition of a document. By default new editions do not have a change note and are therefore invalid.

Ideally we would have set up some validation contexts for the submit and publish actions, but this wasn't possible without changing the way the edition services work substantially. The edition services often validate the edition before they perform the state transition, so it is difficult to enable the change note validation only when we are about to move the edition into a submitted or published state.

We had to remove the Edition#save_as method because it did not support passing a context when saving the edition. We chose to override the existing save method, adding an option for setting the current user, instead of copying the save parameters on to the save_as method.

We would have liked to use the except_on validation context option to enable change note validation for all contexts other than save_draft, but it will not be available until we upgrade to Rails 8.

Added this behind a feature flag, so that we can revert if the slight change in user behaviour causes user dissatisfaction.