Stop validating change notes for draft editions
We only want to make sure that editions have a change note at the point that they are submitted for publication (or force published/scheduled).

Validating change notes for draft editions could cause bugs. For example, if the user replaced an attachment before adding a change note to the edition, the replacement silently failed, leaving the old attachment live.

We therefore change the validation logic to only validate non-draft editions.

Unfortunately this required us to temporarily update the state of the editions in the edition publisher and scheduler services. Those services trigger the edition validation before the state is updated, so the validation occurs as if the editions are still drafts, which isn't logical behaviour.