cancel rescheduling
with other forms, the schedule publishing page
appears after a draft Edition has been created, so
if a user abandons the form we want to delete
the draft Edition.
With rescheduling, we don't want to delete the whole
Edition, because the user is just editing the scheduling of
an already scheduled Edition. So we just return to the
Document view.