Don’t record drafts in version history
When scheduling a content block, we see an additional entry in the
version history to say a draft has been created. This is caused by
the fact that we save the edition before scheduling, and then call
`schedule!`, which triggers another event. It’s tricky to get round
this double save issue as we need to dequeue any previously scheduled
editions as well as update the document reference, so the easiest way
to do this is to skip creating a version when a draft is updated.

I’ve also made some tweaks to the tests so we can be sure the number
of versions is changed by the change to a model’s state.