Only update attachments that have never been published before when publishing and include parent document urls
We want to ensure that the draft value for each asset is set to false when the attachable is published, but Asset Manager does not support updates to assets that are already deleted and not drafts. There's also no real need to update any asset that is not a draft in Asset Manager when publishing. The only time an update might be required is after editing a slug in Whitehall, but that should be handled following the slug editing operation (however it isn't currently). Note that we may still need to delete assets that are not drafts, so we always attempt to delete any assets where attachment_data.deleted? is true.

The only way we can think of identifying which attachment data objects need publishing is to check if the attachment data has a single publicly visible attachable, which would imply that the attachment data was created during the editing of the just published edition.

The previous test was inaccurate in its setup in that it assumed the PublishAttachmentAssetJob would operate with a draft attachable, but in reality it is only enqueued after the state of the attachable has been set to published in the database.

We also include the parent document URL for every asset, to ensure that updating a deleted asset for an attachable that has just been published sets a public Asset Manager URL instead of the draft-origin URL. Asset Manager validates that non-draft assets do not have a draft-origin URL.