Create new edition when updating a schedule
This ensures we flow through the steps correctly and keep state without
having to pass params. To do this, we make a copy of the latest edition,
apply the new schedule/publish params, and redirect to the review step.

I’ve also added a test to ensure we can delete the newly created edition
if we choose to cancel on the confirm page.