Refactor the workflow steps
This tidies up some of the logic and puts the show/update methods in
their own concerns, so it’s clear what they relate to. Also, we 404 if
an unknown step is provided.

I’ve also removed the `*_STEPS` constants, as it wasn’t clear what value
they were offering, especially as now we 404 when an unknown step
is provided.