Remove unused locales
As part of investigating requirements for adding a new language
to Whitehall, we came across these locale files and concluded that
they're not required.

1. Whitehall does not have an option for changing the locale of
   its UI.
2. Even if it did, the majority of the UI uses hardcoded English
   as opposed to i18n.
3. Even if it used i18n, a lot of the translation keys in non
   English locales have been left blank.

We suspect these are a throwback to the days of Whitehall Frontend,
or an optimistic view that Whitehall should fully support i18n and
that we have to start somewhere. But until we take a proper look
at i18n as a whole, this halfway house solution causes confusion
for no benefit.

Trello: https://trello.com/c/T2neQSyf