add a field diff data object
a list of these will be saved against the json column
`field_diffs` on a version. Because of the way these
need to be saved and iterated through (e.g. for one thing it's not a normal one to many relationship, we won't want to add to the list ever again), it's much easier and makes more sense for the column to remain json and we define the
data here.