Remove remaining i18n-dependent tests
These tests were asserting that every locale file has a given
property and that that property is being used by the code.
Now that the locale files themselves have gone away, there's no
value in retaining these tests (which are, of course, failing).

As far as I can gather, these properties were added 12 years ago
when Whitehall Frontend was very much a thing, and thus these
values would be rendered in the frontend.
https://github.com/alphagov/whitehall/commit/c527d38a548ddfed57c24eed2a6fd7c3e0bb6962

They were updated as recently as September 2022, which I believe
also predates the migration of world location news pages away
from Whitehall Frontend:
https://github.com/alphagov/whitehall/pull/6844

Whitehall Frontend is no longer used, and we're left with a
legacy system of locale files and translations when in fact, the
non-English translations of these properties are never used
anywhere.