Move some retagging tests down to the controller level
As suggested in https://github.com/alphagov/whitehall/pull/9834/commits/1253635b349463c267e600bf5f505abcc51dc3a0#r1928517553,
Cucumber tests are expensive and so if we can move the tests
that aren't covering the happy path, into a faster, lower level
area (i.e. controller tests), then that's a good idea.

The dry run with errors test has been moved. The dry run and
successful preview test has been removed, as that's already
tested by the 'non-dry' run Cucumber test.

For the test that has been removed, we've unfortunately lost one
assertion: that the <textarea> contains the value we've POSTed to
the controller. I tried every variation of the following, but
it looks as though `@response.body` is `""` when we `render :index`,
even if we add `layout: false` etc. Somewhere in our Rails test
configuration we must be disabling redirected renders or something.

```ruby
assert_select "textarea[name='csv_input']", csv_to_submit.strip
```

In practice, it's not a huge issue that we're missing the test, as
it isn't critical functionality: publishers won't be filling in
this field by hand, they'll be copying and pasting from another
source.