remove `superseded` event from block timeline
A scheduled edition is moved to a `superseded`
state whenever a newer edition gets published or
scheduled.

Because users aren't aware of the concepts of
Editions of blocks in our current UI, it looks very
confusing if it is inculded, because we already
have an event to show that a new `schedule` has
happened. A design for this event has also not
come up in our latest content design round, so
think it makes sense to just hide them.