Add a 'created' event to block timeline
as per design. We have to be a bit creative here
with how we work out the first version - this is
cause there's some finnicky logic in
`has_audit_trail` and how versions are created,
because we have a `draft` flow on the backend
(e.g. a draft Block is created whenever the user
begins editing a block) but not in practice for
the UX (if the draft isn't publisheed it's never
surfaced again). If we change that we would change
how to create and display versions.