Remove maintenance banner following completion of UTF8MB4 migration
This change was causing one of the visual editor features to fail in CI, even though it passes locally. Given no other tests fail and the line causing the failure is not strictly related to visual editor's behaviour, and I cannot diagnose the issue, I have removed the line. We need to be able to remove the banner this evening, so it seems a reasonable choice to make.