Allow embedded objects to be returned in schema
This allows us to pull out an embedded schema from any properties that
have a `type` of `object`. As we’re making some assumptions about the
shape of a hash, we raise an error if the shape is not what we expect.

Additionally, we remove any embedded objects from the `fields` method.