Ensure assets are deleted when policy group is deleted
Since Policy Groups are not editionable, their delete flows are different. Because of this, when deleting a group, the assets associated with their attachments did not get deleted in Asset Manager.

These changes address that bug, and ensure we delete both the attachments themselves and their corresponding assets in Asset Manager. Rather than trying to reuse some of the classes that serve the edition workflow, we preferred to use a callback on the PolicyGroup model, only reusing the asset specific DeleteAttachmentAssetJob.