Only check `safely_resluggable?` before generating new ID
See code review comment:
https://github.com/alphagov/whitehall/pull/9892#discussion_r1942515565

`safely_resluggable?` returns `true` for newly created attachments
and `false` for attachments that have been copied over from a
previous edition.

New attachments will start off with a `nil` slug, or in any case
a slug that is derived from the attachment's current title. It
can be safely re-generated an infinite number of times before the
attachment is made live. So we don't need to be checking the
value of the slug in this method.