WIP add search index method to landing page
This is needed because the Attachable module assumes that the class its
attached to has a search_index method, which landing page doesn't. This
usually comes from the Searchable module, but we don't need any of that
behaviour.

Actually we don't need Attachable either, I don't think - I think it's a
hangover from when we were attaching CSVs to use for line graphs. But
we're not doing line graphs now. We could remove it, but I'm slightly
nervous about breaking something.