Fix date input name bug
Rails nested form inputs should be in the style "parent[level_one_name][level_two_name]", but forms also used to work in the style "parent[level_one_name[level_two_name]]". As of Rails 8 that is no longer the case, so we need to correct the form input names.