Embedded objects aren’t removed when editing
Before, when editing an object we were creating an edition from scratch
with only the params given. This means that when we have embedded
objects (that aren’t present in the form), they get removed.

This updates the create edition service to merged the details from the
previous edition with the params to ensure nothing is removed.