Populate document_type using Rails variables
Builds upon the exploration of using controller_name as the
document type in #9593 to use an interpolated string of
both action_name and controller_name as an imperfect proxy
for the document_type.

PAs have a need for the document_type to be provided because
this will mirror the implementation on GOV.UK. Because there
is no concept of a document_type in Whitehall, PAs are content
for this to act as a proxy.