Fix rendering embed codes in preview host documents
The helper method in Content Block Tools that finds
content block embeds was recently set to
return mutliple references to one block [1] - this
then caused a knock-on effect for rendering blocks in
a document, because it would try to replace the
embed code multiple times, causing issues now that
the embed code is returned in the data attributes of
a span. We should only need to run gsub once to replace
all the embed codes for a reference, so I'm calling `uniq`
in the code that uses the Tool.

[1] alphagov/govuk_content_block_tools#6