fix custom setting of header margin
This was changed recently in this PR [1] to set a
default margin of 8, however it removed the
setting of custom margins via the
`title_margin_bottom` flag. So anywhere that had
set that was being ignored. This reinstates the
check for a custom setting, and if it isn't
present adds the default `8`.

[1] https://github.com/alphagov/whitehall/pull/9863