Add a `generate_diff` method to a concern
This will soon be used to replace the FieldDiff code. Rather than
returning an array of FieldDiff objects, this returns a keyed hash
of fields and an array with the previous value and current value,
inspired by the [Changesets in the PaperTrail gem][1]

This will make it easier to fetch deeply nested fields when showing the
version history.

[1]: https://github.com/paper-trail-gem/paper_trail?tab=readme-ov-file#3c-diffing-versions