Add new 'include_blank' option for autocomplete module
It's a common use case to show a blank option on the select
dropdown, and it feels wrong that in each case where that's
required, we have to prefix options with `[""]`. It should just
be a natively supported feature (allowing us to change the
underlying implementation much more easily).

This commit carries over the `include_blank` option from the
'select_with_search' component so that we have feature parity
between the two components.

It also uses the new option on existing 'autocomplete'
instances.