Set showAllValues to 'true' whether selecting multiple or single
In March 2023, in #7378, it was decided to make 'multiple select'
autocomplete occurrences use the `showAllValues` option by default.
This means that all options are displayed as soon as the input is
interacted with, without any additional search action required.

The PR did not describe why 'single' inputs should not receive the
same treatment. It is hypothesised that without showing users the
available options (i.e. enabling `showAllValues`) then they will
find it more difficult to know which value to search for and select.

If we don't change this behaviour, then it means we can't offer
feature parity in swapping component usage from 'select with search'
over to 'autocomplete' (which is what we're trying to do in
https://trello.com/c/pmbL2yId/3411-remove-the-selectwithsearch-component-from-whitehall
to improve consistency within Whitehall).

Note that the behaviour can still be overridden on a case by case
basis as per the example in the YML file:

```yml
      autocomplete_configuration_options:
        showAllValues: false
```