Don’t change the object key when changing the name
This avoids the issue where changing the name of a rate breaks the embed
code. It does mean that once a name is set, this means the key of the
object is immutable, but this is a similar issue to how slugs work in
other documents. We can revisit this if it becomes an issue.