Add a no javascript option for embed codes in embedded objects
This adds a row for each field showing the embed
code. The row will be removed if javascript is
enabled in the `copy-embed-code` script. It isn't
an ideal design at the moment, but as we believe
it to be an edge case for an internal tool we
think it will be usable for now.